Remove disconnect() method
The method was not working because it was searching the connections with
the hostname, but we use IP addresses.  It is a good idea to disconnect
after some operations, but I would bother fixing it.  Currently we are
calling disconnect_all() after we are done, anyway.  We aim to get rid
of Fabric.  Less code we have better it is.