Allow for run() without sudo
We want to run all commands via sudo apart from one: virsh migrate.
This command runs from non-root account on one HV and connects to the
same account on another one. Should it run via sudo on the 1st one, it
would not have keys to connect to the other one.