Merge hypervisor constraints and rules
The previous "balance" submodule designed with separate constraints and
rules.  The constraints were causing hypervisors to be filtered out, and
the rules were causing them to be ordered.  Filtering out behavior isn't
really what we want.  Such hypervisors can still be used, when we don't
have anything else.  Letting them be selected helps, at least, the test
suite.

On the previous design, the return values of the rules were combined
into each other, but the current implementation of the rules were
returning such results that some were always more important that others.
This change removes the possibility of preferences being equally
important.  This simplification on the design would let us troubleshoot,
assess, and fine tune the process much easier.

With this change, we also clearly don't need to keep the "balance" as
a namespace.  The net lines of code deleted until this commit is close
to 80%.  On the new design there is a brief class only to do ranking of
the hypervisors and the preferences to hold the former constraints and
rules together.  It is better not to use "balance" on the names of those
two new modules to reserve it to the module which we will automatically
"balance" our hypervisor load hopefully one day.