Handle hypervisor preferences returning None
After getting the refactoring in, I noticed that the previous fix
wasn't the good one.  First of all it was not the correct place.
LazyCompare class should only provide lazy comparison.  It has nothing
to do with Nones.  Also not all hypervisor preferences can return None,
so LazyCompare would do some unnecessary for work for them.
Furthermore some preferences need to handle Nones before returning
a result.  The previous fix was not covering them.  This one at least
cover Nones on HypervisorAttributeValueLimit as well as
HypervisorAttributeValue.