Improve tests
Non-conflicing function names are used. This does not seem a real issue but is
easier to understand the code for human beings.

All tests are happening on automatically selected hypervisors. That
means there are a few tests less now, as there are no test for
pre-selected hypervisor.

When contacting a VM or its HV is necessary, a fresh object is queried each
time.