lock: Make igvm_locked a datetime, not boolean
Sometimes hypervisors and VMs are not porperly unlocked when igvm
crashes, the process get killed, the terminal closed, ctrl-c pressed
many times etc..

To make it easier to monitor this problem we are changing igvm_locked to
a datetime.  This way it becomes easy to find all objects locked longer
then X via a simple serveradmin query.