Fix host objects lock not released after crash
We experienced that some objects such as hypervisors and VMs are not
porperly release their lock when igvm crashed or gets aborted by closing
the terminal or hitting ctrl-c many times in a row or the process gets
killed etc.

To get rid of this problem we should change igvm_locked to a number value
and work with timestamps so that we can check if a object has been locked
for too long as it is currently not possible to query serveradmin
history via API and this requires greater refactoring.