Fix the used hypervisor preferences log message
On the previous log message, the number was one off, because the loop
is broken after hitting the not-executed preference.  The information
was also wrong, because it doesn't really mean that many preferences
are used to select this particular hypervisor.  It may be that sorting
this hypervisor before the next one caused the preferences to be
executed.  We can do better in here by comparing the number of executed
preferences for the next hypervisor on the line, or rolling out our own
sorting algorithm to be able to track this information, but I don't
think the information is worth the effort.