Remove the Jenkins build number from testing hostnames
Since the Jenkins executor id is already present in the hostnames, the
build number does not add any value. Instead it is much easier now to
clean the right things in setUp and tearDown and there is no need to
clean really everything (from previous builds) on the HVs anymore.