Fix an issue where exceptions are lost
The actual fix is moving the "yield" out of the loop.
This prevents a StopIteration exception from being raised when the
generator is interrupted. When a StopIteration exception is raised,
the ExitStack tries to do some black magic and in fact drops the
exception together with any original exception that was passed
to the ExitStack.
By getting rid of that code path, the exception can be properly passed
to any outside caller and handle it properly.