Fix test_vm_rename test has left over VMs
With the original fix of the vm_rename command and its newly added tests
we create VM with a hostname pattern not covered by the cleanup task.
This lead to left over VMs breaking other test runs.

We now rename the VM back to its original test in the vm_rename test. In
case the test fails we leave it up to a human to delete the VM.