Improve handling of AWS overview file (#184)
* Make AWS instances overview editable for all users in adm

The owner of the etag and json overview file is per
default the user, who downloaded the file. We need
to allow all users, who are using igvm, to overwrite
these files, if needed (e.g. due to an update).

* Open AWS instances overview file for read and write

We need to open the overview file with r+/w+ to
eliminate a race condition while using igvm in
parallel with a huge amount of VMs.

Co-authored-by: Christian Matthaei <christian.matthaei@innogames.com>