Add temp workaround for faulty estimate_cpu_usage
Apparently Hypervisor.estimate_cpu_usage() can be negative if the
migration log sums up to a big enough negative value. It would then
outweigh the calculated HV CPU usage and the estimated VM CPU usage
on that HV.
The workaround is to correct the sum to be at least zero for now.
Fixing the root cause will be handled in another PR.