Add a flag to overrule Hypervisor preferences
This adds a flag "--soft-preferences" to the build, migrate and
evacuate commands of igvm. Setting them causes the Hypervisor
selection algorithm to behave slightly different. Instead of
excluding Hypervisors if any of the preferences fails it will instead
keep them with an accordingly lower ranked overall score. This way we
can force all Hypervisors to be considered as targets while still
leaving the hard checks for later on to check if VMs really fit.
In emergency cases this can help emptying Hypervisors or just moving
a specific VM to the "less worst" available location. It can also be
used to force a VM build and squeeze it in somewhere.