Remove cpu_util_vm_pct preference from old times
We used to express the CPU usage limit of HVs with the
cpu_util_vm_pct metric. For quite some time now, though, we came up
with a better way by setting specific thresholds for specific
hardware models that are backed by load tests. There is no reason to
keep the old limiting preference anymore. We can safely drop it.