Fix checks for VM/storage existence
A new committed object does not carry an object_id and so it is None. This None
value is inserted into a string and that is used to check for VM/storage
existence in the tests.
As a result of this all checks for non-existence have always passed.
This change refreshes the object from SA and therefore produces correct
existence check commands.