Allow specifying the target HV as a query
Previously it was only possible to specify specific single target HV
hostnames. This change allows specifying a Serveradmin query instead
to select a range of possible HVs to pick from. The "best" HV among
those will automatically be chosen by igvm.
This affects the "migrate" and "evacuate" commands and also adds this
parameter to the "build" command.
The changes are fully backwards-compatible and can be used just like
before, resulting in the same behavior.