Fix xfs offline transfer for bullseye
Some software in bullseye got a bit more picky with handling default
file handles, like stderr. I suppose the behavior is actually correct
now on bullseye.
Previously the command would return just fine and detach to the
background and the ssh channel would close. However, we never
redirected stderr of the nc.openbsd command, so it was still attached
to the ssh channel. On bullseye this will now keep the channel open.
This leads to the command never returning and blocking the execution
flow forever, resulting in a stuck migration.
The fix is to add a stderr redirection to the nc.openbsd command.