Fix igvm does not specify minimum dependencies
The install_requires via setup.py is the proper way to specify the bare
minimum dependencies to be able to run a Python app. requirements.txt is
to specify an environment.

However for igvm it is the same so we parse the requirements.txt so that
source dists and wheels have dependencies properly specified.

We might replace this at a later point with pyprojects.toml