Pin markupsafe dependency to 1.1
Markupsafe 1.1.1 is the latest version shipped with Debian Bullseye
which we are currently using.

Not specifying a version makes us run into a bug where markupsafe
removed a dependency in a minor version (semver violation) when using
jinja2 in version 2.x.

See: https://github.com/pallets/markupsafe/issues/304

This can be removed once we are on Debian Bookworm using Jinja 3.x