Repin all possible vCPUs, not just ones in use
When trying to increse the amount of vCPUs in a VM, we were greeted with
a cgroup error:
```
 virCgroupSetValueRaw:502 : Unable to write to '/sys/fs/cgroup/machine.slice/.../vcpu6/cpuset.cpus': Numerical result out of range
```

It turns out that when live repinning the VM on the new hypervisor, we
were iterating only over the amount of vCPUs assigned to the VM (for
example, 0 to 5) instead of the max amount of vCPUs assignable (coming
from props.max_cpus).

Now we iterate over the max amount of vCPUs assignable and pin them to
the local hypervisor topology.