Use new auto assign logic for ip addresses
We no longer want users to pick ip addresses but just specify the
network and then they get automatically an ip address. The exception is
manually setting an ip address in which case we can use the adminapi to
get an ip address and lock it so others do not use it.