Use modern machine type and emulator on domain XML
RHEL is deprecating the i440 machine type in favor of the q35.

This patch switches the machine type to q35 and updates the root
controller to be PCIe instead of PCI (a requirement of q35).

We also update the path to the emulator to use the qemu binary directly,
instead of the kvm symlink that points to qemu.